Return nil when openFifo returns nil
The return statement in the OpenFifo wrapper function implicitly boxes
the *fifo pointer into an io.ReadWriteCloser interface value, even when
the value is (*fifo)(nil). It is the same gotcha described in
https://go.dev/doc/faq#nil_error and causes the same sorts of confusing
situations. Modify the OpenFifo wrapper function to return a nil
interface value when openFifo returns a nil pointer value.

Signed-off-by: Cory Snider <csnider@mirantis.com>